Added Bounds and player rotation
Added a boundary to the map in the form of a square areana that blocks the player in. Still need to setup the arena to stop the player as player can currently fly through the box.

Added player rotation so that as they move they rotate towards the direction in which they are moving.